calc validation accuracy during training
In this commit the calculation and printing of the validation
accuracy during training was implemented. In order to do that we
altered the `getBestThreshold()` function in `Test.h` so that it
now returns the validation accuracy as well.

Of course, changes in `Config.py` had to happen as well. A new
function named `validation_acc()` that calculated the validation
accuracy was created. We also refartored some other things such as
when the validation batch is read (now in the beginning of training
).

The train_test.ipynb was also adapted.