fix value of feat_dicts in parse_feature_matrix
The function `parse_feature_matrix` was moved to
`linear_regression_analysis.py` and it was changed in order to correct
the value of each feature in `feat_dicts`.

Before the change, a string was being passed to the values, which caused
`DictVectorizer` to misinterpret the feature names.

Another correction was to `rstrip()` the line string when reading the
feature matrix files in order to remove the newline (`\n`)