script to extract features (outside notebook)
A python script was created to extract features using PRA
algorithm. The goal of this script is to unify feature extraction
for different cases, such as using ids or not and from g_hat or
not.

There are still lots of hardcoded paths and files, this should be
changed in the future.